CA Cambio de estructura de proyecto para dividir ACTIVIDADES, MODELOS, CLASE DATABASE
