+++ ANDROID/Proyecto App Mobil/app/src/main/java/com/example/virtualtrendsmovile/actividades/ConfirmacionFinal.java
@@ -1,38 +1,38 @@
-package com.example.virtualtrendsmovile;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
+package com.example.virtualtrendsmovile.actividades;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.virtualtrendsmovile.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
 
-public class ComprobacionReserva extends AppCompatActivity {
+public class ConfirmacionFinal extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_comprobacion_rerserva);
+        setContentView(R.layout.activity_confirmacion_final);
         BottomNavigationView nav = findViewById(R.id.btnNavSelector);
         nav.setSelected(true);
-
 
         nav.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
 
                 int id = item.getItemId();
-                if (id == R.id.back) {
-                    startActivity(new Intent(getApplicationContext(), ConfirmarDatosPersonales.class));
-                } else if (id == R.id.info) {
+                if(id==R.id.back){
+                    startActivity(new Intent(getApplicationContext(), ComprobacionReserva.class));
+                } else if (id==R.id.info) {
                     startActivity(new Intent(getApplicationContext(), ContactoActivity.class));
-                } else if (id == R.id.map) {
+                }else if (id== R.id.map){
                     startActivity(new Intent(getApplicationContext(), DondeEstamosActivity.class));
-                } else if (id == R.id.turn) {
+                } else if (id==R.id.turn) {
                     startActivity(new Intent(getApplicationContext(), NuestroServicio.class));
-                } else if (id == R.id.logout) {
+                } else if (id==R.id.logout) {
                     startActivity(new Intent(getApplicationContext(), InicioActivity.class));
                 }
 
@@ -40,9 +40,11 @@ public class ComprobacionReserva extends AppCompatActivity {
             }
 
         });
+
     }
-    public void comprobacion_reserva(View view){
-        Intent intent = new Intent(this, ConfirmacionFinal.class);
+
+    public void confirmar_final_reserva(View view){
+        Intent intent = new Intent(this, ComprobacionReserva.class);
         startActivity(intent);
         finish();
     }
