+++ ANDROID/Proyecto App Mobil/app/src/main/java/com/example/virtualtrendsmovile/actividades/SelectorActivity.java
@@ -1,32 +1,46 @@
-package com.example.virtualtrendsmovile;
+package com.example.virtualtrendsmovile.actividades;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.Button;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
+import com.example.virtualtrendsmovile.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
 
-public class TurnosActivity extends AppCompatActivity {
 
-
+public class SelectorActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_turnos);
+        setContentView(R.layout.activity_selector);
+
+
+        Button botonIrANuestroServicio = findViewById(R.id.botonIrANuestroServicio);
+
+        botonIrANuestroServicio.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // Iniciar la actividad NuestroServicio cuando se haga clic en el botón.
+                Intent intent = new Intent(SelectorActivity.this, NuestroServicio.class);
+                startActivity(intent);
+            }
+        });
+
         BottomNavigationView nav = findViewById(R.id.btnNavSelector);
-        nav.setSelected(true);
+        nav.setSelectedItemId(R.id.back);
         nav.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
                 int id = item.getItemId();
                 if(id==R.id.back){
-                    startActivity(new Intent(getApplicationContext(), SelectorActivity.class));
+                    startActivity(new Intent(getApplicationContext(), LoginActivity.class));
                 } else if (id==R.id.info) {
                     startActivity(new Intent(getApplicationContext(), ContactoActivity.class));
                 }else if (id== R.id.map){
@@ -39,13 +53,13 @@ public class TurnosActivity extends AppCompatActivity {
                 return false;
             }
         });
-
     }
-    public void ejecutar_turno(View view){
-        Intent intent = new Intent(this, TurneroActivity.class);
+
+
+    public void pantalla_turnos(View view){
+        Intent intent = new Intent(this, TurnosActivity.class);
         startActivity(intent);
         finish();
     }
-
 
 }
