+++ ANDROID/Proyecto App Mobil/app/src/main/java/com/example/virtualtrendsmovile/actividades/TurneroActivity.java
@@ -1,4 +1,4 @@
-package com.example.virtualtrendsmovile;
+package com.example.virtualtrendsmovile.actividades;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -8,23 +8,24 @@ import android.os.Bundle;
 import android.view.MenuItem;
 import android.view.View;
 
+import com.example.virtualtrendsmovile.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
 
-public class ConfirmarDatosPersonales extends AppCompatActivity {
+public class TurneroActivity extends AppCompatActivity {
+
+    @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_confirmar_datos_personales);
+        setContentView(R.layout.activity_turnero);
         BottomNavigationView nav = findViewById(R.id.btnNavSelector);
         nav.setSelected(true);
-
         nav.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-
                 int id = item.getItemId();
                 if(id==R.id.back){
-                    startActivity(new Intent(getApplicationContext(), HorariosTurnosActivity.class));
+                    startActivity(new Intent(getApplicationContext(), TurnosActivity.class));
                 } else if (id==R.id.info) {
                     startActivity(new Intent(getApplicationContext(), ContactoActivity.class));
                 }else if (id== R.id.map){
@@ -34,18 +35,14 @@ public class ConfirmarDatosPersonales extends AppCompatActivity {
                 } else if (id==R.id.logout) {
                     startActivity(new Intent(getApplicationContext(), InicioActivity.class));
                 }
-
                 return false;
             }
-
         });
 
     }
-
-    public void confirmar_datos(View view){
-        Intent intent = new Intent(this, ComprobacionReserva.class);
+    public void ejecutar_turno(View view){
+        Intent intent = new Intent(this, HorariosTurnosActivity.class);
         startActivity(intent);
         finish();
     }
-
 }
