+++ ANDROID/Proyecto App Mobil/app/src/main/java/com/example/virtualtrendsmovile/actividades/TurnosActivity.java
@@ -1,33 +1,33 @@
-package com.example.virtualtrendsmovile;
+package com.example.virtualtrendsmovile.actividades;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.view.MenuItem;
+import android.view.View;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.view.MenuItem;
-import android.view.View;
-
+import com.example.virtualtrendsmovile.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
-import com.google.android.material.snackbar.Snackbar;
 
-public class ContactoActivity extends AppCompatActivity {
+public class TurnosActivity extends AppCompatActivity {
+
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_contacto);
-
+        setContentView(R.layout.activity_turnos);
         BottomNavigationView nav = findViewById(R.id.btnNavSelector);
-        nav.setSelectedItemId(R.id.info);
+        nav.setSelected(true);
         nav.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
             @Override
             public boolean onNavigationItemSelected(@NonNull MenuItem item) {
                 int id = item.getItemId();
                 if(id==R.id.back){
-                    startActivity(new Intent(getApplicationContext(), LoginActivity.class));
+                    startActivity(new Intent(getApplicationContext(), SelectorActivity.class));
                 } else if (id==R.id.info) {
                     startActivity(new Intent(getApplicationContext(), ContactoActivity.class));
                 }else if (id== R.id.map){
@@ -40,19 +40,13 @@ public class ContactoActivity extends AppCompatActivity {
                 return false;
             }
         });
+
+    }
+    public void ejecutar_turno(View view){
+        Intent intent = new Intent(this, TurneroActivity.class);
+        startActivity(intent);
+        finish();
     }
 
-    public void enviar(View view) {
-
-        Snackbar.make(view, "Mensaje recibido", Snackbar.LENGTH_SHORT).show();
-
-        view.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent = new Intent(ContactoActivity.this, InicioActivity.class);
-                startActivity(intent);
-            }
-        }, 1500);
-    }
 
 }
