+++ ANDROID/Proyecto App Mobil/app/src/main/java/com/example/virtualtrendsmovile/actividades/pageUsuarios.java
@@ -1,0 +1,4 @@
+package com.example.virtualtrendsmovile.actividades;
+
+public class pageUsuarios {
+}
