--- ANDROID/Proyecto App Mobil/app/src/main/java/com/example/virtualtrendsmovile/pageUsuarios.java
@@ -1,4 +1,0 @@
-package com.example.virtualtrendsmovile;
-
-public class pageUsuarios {
-}
